requirements.txt
GitPython==3.1.34
ass==0.5.4